refactor: removed object_id module
refactor: relocated its function to common/validation_util
