+++ src/apps/backend/modules/account/internal/store/account_model.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, ConfigDict, Field
 from pydantic.functional_validators import AfterValidator
 
 from modules.account.types import PhoneNumber
-from modules.object_id.utils import object_id_validate
+from modules.common.validation_utils import object_id_validate
 
 PyObjectId = Annotated[ObjectId | str, AfterValidator(object_id_validate)]
